+++ pallets/automation-time/src/tests.rs
@@ -19,7 +19,7 @@ use crate::{
 	migrations::{v1, v2},
 	mock::*,
 	Action, Error, LastTimeSlot, MissedQueue, MissedTask, Shutdown, Task, TaskHashInput, TaskQueue,
-	Tasks, WeightInfo,
+	Tasks, WeightInfo
 };
 use core::convert::TryInto;
 use frame_support::{assert_noop, assert_ok, error::BadOrigin, traits::OnInitialize};
@@ -40,7 +40,7 @@ fn schedule_invalid_time() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME + 1],
 				vec![12]
@@ -55,7 +55,7 @@ fn schedule_past_time() {
 	new_test_ext(START_BLOCK_TIME + 1_000 * 10800).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME],
 				vec![12]
@@ -65,7 +65,7 @@ fn schedule_past_time() {
 
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME - 3600],
 				vec![12]
@@ -80,7 +80,7 @@ fn schedule_too_far_out() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME + 1 * 24 * 60 * 60],
 				vec![12]
@@ -95,7 +95,7 @@ fn schedule_no_message() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME],
 				vec![]
@@ -110,7 +110,7 @@ fn schedule_no_provided_id() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![],
 				vec![SCHEDULED_TIME],
 				vec![12]
@@ -125,7 +125,7 @@ fn schedule_not_enough_for_fees() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![60],
 				vec![SCHEDULED_TIME],
 				vec![12]
@@ -138,10 +138,10 @@ fn schedule_not_enough_for_fees() {
 #[test]
 fn schedule_notify_works() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		let message: Vec<u8> = vec![2, 4, 5];
 		assert_ok!(AutomationTime::schedule_notify_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			vec![50],
 			vec![SCHEDULED_TIME],
 			message.clone()
@@ -156,7 +156,7 @@ fn schedule_notify_works() {
 				},
 				Some(task) => {
 					let expected_task = Task::<Test>::create_event_task(
-						AccountId32::new([ALICE; 32]),
+						AccountId32::new(ALICE),
 						vec![50],
 						vec![SCHEDULED_TIME].try_into().unwrap(),
 						message,
@@ -174,10 +174,10 @@ fn schedule_native_transfer_invalid_amount() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_native_transfer_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME],
-				AccountId32::new([BOB; 32]),
+				AccountId32::new(BOB),
 				0,
 			),
 			Error::<Test>::InvalidAmount,
@@ -190,10 +190,10 @@ fn schedule_native_transfer_cannot_transfer_to_self() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_native_transfer_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME],
-				AccountId32::new([ALICE; 32]),
+				AccountId32::new(ALICE),
 				1,
 			),
 			Error::<Test>::TransferToSelf,
@@ -204,12 +204,12 @@ fn schedule_native_transfer_cannot_transfer_to_self() {
 #[test]
 fn schedule_native_transfer_works() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		assert_ok!(AutomationTime::schedule_native_transfer_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			vec![50],
 			vec![SCHEDULED_TIME],
-			AccountId32::new([BOB; 32]),
+			AccountId32::new(BOB),
 			1,
 		));
 		match AutomationTime::get_scheduled_tasks(SCHEDULED_TIME) {
@@ -222,10 +222,10 @@ fn schedule_native_transfer_works() {
 				},
 				Some(task) => {
 					let expected_task = Task::<Test>::create_native_transfer_task(
-						AccountId32::new([ALICE; 32]),
+						AccountId32::new(ALICE),
 						vec![50],
 						vec![SCHEDULED_TIME].try_into().unwrap(),
-						AccountId32::new([BOB; 32]),
+						AccountId32::new(BOB),
 						1,
 					);
 
@@ -297,7 +297,7 @@ fn schedule_xcmp_errors_bad_origin() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		assert_noop!(
 			AutomationTime::schedule_xcmp_task(
-				cumulus_pallet_xcm::Origin::SiblingParachain(PARA_ID.try_into().unwrap()).into(),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME],
 				PARA_ID.try_into().unwrap(),
@@ -312,16 +312,16 @@ fn schedule_xcmp_errors_bad_origin() {
 #[test]
 fn schedule_duplicates_errors() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		assert_ok!(AutomationTime::schedule_notify_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			vec![50],
 			vec![SCHEDULED_TIME],
 			vec![2, 4, 5]
 		),);
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![SCHEDULED_TIME],
 				vec![2, 4]
@@ -334,10 +334,10 @@ fn schedule_duplicates_errors() {
 #[test]
 fn schedule_max_execution_times_errors() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![50],
 				vec![
 					SCHEDULED_TIME,
@@ -355,7 +355,7 @@ fn schedule_max_execution_times_errors() {
 #[test]
 fn schedule_execution_times_removes_dupes() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		let task_id1 = schedule_task(
 			ALICE,
 			vec![50],
@@ -374,7 +374,7 @@ fn schedule_execution_times_removes_dupes() {
 			},
 			Some(task) => {
 				let expected_task = Task::<Test>::create_event_task(
-					AccountId32::new([ALICE; 32]),
+					AccountId32::new(ALICE),
 					vec![50],
 					vec![SCHEDULED_TIME, SCHEDULED_TIME + 10800].try_into().unwrap(),
 					vec![2, 4],
@@ -389,15 +389,15 @@ fn schedule_execution_times_removes_dupes() {
 #[test]
 fn schedule_time_slot_full() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		assert_ok!(AutomationTime::schedule_notify_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			vec![50],
 			vec![SCHEDULED_TIME],
 			vec![2, 4]
 		));
 		assert_ok!(AutomationTime::schedule_notify_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			vec![60],
 			vec![SCHEDULED_TIME],
 			vec![2, 4, 5]
@@ -405,7 +405,7 @@ fn schedule_time_slot_full() {
 
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![70],
 				vec![SCHEDULED_TIME],
 				vec![2]
@@ -418,13 +418,13 @@ fn schedule_time_slot_full() {
 #[test]
 fn schedule_time_slot_full_rolls_back() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
 		let task_id1 = schedule_task(ALICE, vec![40], vec![SCHEDULED_TIME + 7200], vec![2, 4, 5]);
 		let task_id2 = schedule_task(ALICE, vec![50], vec![SCHEDULED_TIME + 7200], vec![2, 4]);
 
 		assert_noop!(
 			AutomationTime::schedule_notify_task(
-				Origin::signed(AccountId32::new([ALICE; 32])),
+				Origin::signed(AccountId32::new(ALICE)),
 				vec![70],
 				vec![SCHEDULED_TIME, SCHEDULED_TIME + 3600, SCHEDULED_TIME + 7200],
 				vec![2]
@@ -460,11 +460,11 @@ fn cancel_works_for_scheduled() {
 		System::reset_events();
 
 		assert_ok!(AutomationTime::cancel_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			task_id1,
 		));
 		assert_ok!(AutomationTime::cancel_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			task_id2,
 		));
 
@@ -475,11 +475,11 @@ fn cancel_works_for_scheduled() {
 			events(),
 			[
 				Event::AutomationTime(crate::Event::TaskCancelled {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id1
 				}),
 				Event::AutomationTime(crate::Event::TaskCancelled {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id2
 				}),
 			]
@@ -500,7 +500,7 @@ fn cancel_works_for_multiple_executions_scheduled() {
 		System::reset_events();
 
 		assert_ok!(AutomationTime::cancel_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			task_id1,
 		));
 
@@ -517,7 +517,7 @@ fn cancel_works_for_multiple_executions_scheduled() {
 		assert_eq!(
 			events(),
 			[Event::AutomationTime(crate::Event::TaskCancelled {
-				who: AccountId32::new([ALICE; 32]),
+				who: AccountId32::new(ALICE),
 				task_id: task_id1
 			})]
 		);
@@ -584,7 +584,7 @@ fn cancel_works_for_an_executed_task() {
 		}
 
 		assert_ok!(AutomationTime::cancel_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			task_id1
 		));
 
@@ -595,7 +595,7 @@ fn cancel_works_for_an_executed_task() {
 		assert_eq!(
 			events(),
 			[Event::AutomationTime(crate::Event::TaskCancelled {
-				who: AccountId32::new([ALICE; 32]),
+				who: AccountId32::new(ALICE),
 				task_id: task_id1
 			})]
 		);
@@ -617,14 +617,14 @@ fn cancel_works_for_tasks_in_queue() {
 		assert_eq!(1, AutomationTime::get_task_queue().len());
 
 		assert_ok!(AutomationTime::cancel_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			task_id,
 		));
 
 		assert_eq!(
 			events(),
 			[Event::AutomationTime(crate::Event::TaskCancelled {
-				who: AccountId32::new([ALICE; 32]),
+				who: AccountId32::new(ALICE),
 				task_id
 			}),]
 		);
@@ -638,7 +638,7 @@ fn cancel_must_be_owner() {
 		let task_id = schedule_task(ALICE, vec![40], vec![SCHEDULED_TIME], vec![2, 4, 5]);
 
 		assert_noop!(
-			AutomationTime::cancel_task(Origin::signed(AccountId32::new([BOB; 32])), task_id),
+			AutomationTime::cancel_task(Origin::signed(AccountId32::new(BOB)), task_id),
 			Error::<Test>::NotTaskOwner,
 		);
 	})
@@ -648,7 +648,7 @@ fn cancel_must_be_owner() {
 fn cancel_task_must_exist() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		let task = Task::<Test>::create_event_task(
-			AccountId32::new([ALICE; 32]),
+			AccountId32::new(ALICE),
 			vec![40],
 			vec![SCHEDULED_TIME].try_into().unwrap(),
 			vec![2, 4, 5],
@@ -656,7 +656,7 @@ fn cancel_task_must_exist() {
 		let task_id = BlakeTwo256::hash_of(&task);
 
 		assert_noop!(
-			AutomationTime::cancel_task(Origin::signed(AccountId32::new([ALICE; 32])), task_id),
+			AutomationTime::cancel_task(Origin::signed(AccountId32::new(ALICE)), task_id),
 			Error::<Test>::TaskDoesNotExist,
 		);
 	})
@@ -666,7 +666,7 @@ fn cancel_task_must_exist() {
 fn cancel_task_not_found() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		let task = Task::<Test>::create_event_task(
-			AccountId32::new([ALICE; 32]),
+			AccountId32::new(ALICE),
 			vec![40],
 			vec![SCHEDULED_TIME].try_into().unwrap(),
 			vec![2, 4, 5],
@@ -675,7 +675,7 @@ fn cancel_task_not_found() {
 		<Tasks<Test>>::insert(task_id, task);
 
 		assert_ok!(AutomationTime::cancel_task(
-			Origin::signed(AccountId32::new([ALICE; 32])),
+			Origin::signed(AccountId32::new(ALICE)),
 			task_id,
 		));
 		assert_eq!(
@@ -683,7 +683,7 @@ fn cancel_task_not_found() {
 			[
 				Event::AutomationTime(crate::Event::TaskNotFound { task_id }),
 				Event::AutomationTime(crate::Event::TaskCancelled {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id
 				})
 			]
@@ -702,7 +702,7 @@ fn force_cancel_task_works() {
 		assert_eq!(
 			events(),
 			[Event::AutomationTime(crate::Event::TaskCancelled {
-				who: AccountId32::new([ALICE; 32]),
+				who: AccountId32::new(ALICE),
 				task_id
 			}),]
 		);
@@ -829,22 +829,22 @@ fn trigger_tasks_limits_missed_slots() {
 				[
 					Event::AutomationTime(crate::Event::Notify { message: vec![50] }),
 					Event::AutomationTime(crate::Event::TaskMissed {
-						who: AccountId32::new([ALICE; 32]),
+						who: AccountId32::new(ALICE),
 						task_id: missing_task_id0,
 						execution_time: SCHEDULED_TIME - 25200,
 					}),
 					Event::AutomationTime(crate::Event::TaskMissed {
-						who: AccountId32::new([ALICE; 32]),
+						who: AccountId32::new(ALICE),
 						task_id: missing_task_id5,
 						execution_time: SCHEDULED_TIME - 18000,
 					}),
 					Event::AutomationTime(crate::Event::TaskMissed {
-						who: AccountId32::new([ALICE; 32]),
+						who: AccountId32::new(ALICE),
 						task_id: missing_task_id4,
 						execution_time: SCHEDULED_TIME - 14400,
 					}),
 					Event::AutomationTime(crate::Event::TaskMissed {
-						who: AccountId32::new([ALICE; 32]),
+						who: AccountId32::new(ALICE),
 						task_id: missing_task_id3,
 						execution_time: SCHEDULED_TIME - 10800,
 					}),
@@ -912,7 +912,7 @@ fn trigger_tasks_completes_all_tasks() {
 fn trigger_tasks_handles_nonexisting_tasks() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
 		let task_hash_input =
-			TaskHashInput::<Test>::create_hash_input(AccountId32::new([ALICE; 32]), vec![20]);
+			TaskHashInput::<Test>::create_hash_input(AccountId32::new(ALICE), vec![20]);
 		let bad_task_id = BlakeTwo256::hash_of(&task_hash_input);
 		let mut task_queue = AutomationTime::get_task_queue();
 		task_queue.push(bad_task_id);
@@ -984,12 +984,12 @@ fn trigger_tasks_completes_all_missed_tasks() {
 			events(),
 			[
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id1,
 					execution_time: 0
 				}),
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id2,
 					execution_time: 0
 				}),
@@ -1042,12 +1042,12 @@ fn missed_tasks_updates_executions_left() {
 			events(),
 			[
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id1,
 					execution_time: SCHEDULED_TIME
 				}),
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id2,
 					execution_time: SCHEDULED_TIME
 				}),
@@ -1111,7 +1111,7 @@ fn missed_tasks_removes_completed_tasks() {
 			[
 				Event::AutomationTime(crate::Event::Notify { message: message_one }),
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: task_id01,
 					execution_time: SCHEDULED_TIME
 				}),
@@ -1124,21 +1124,21 @@ fn missed_tasks_removes_completed_tasks() {
 #[test]
 fn trigger_tasks_completes_some_native_transfer_tasks() {
 	new_test_ext(START_BLOCK_TIME).execute_with(|| {
-		get_funds(AccountId32::new([ALICE; 32]));
-		let current_funds = Balances::free_balance(AccountId32::new([ALICE; 32]));
+		get_funds(AccountId32::new(ALICE));
+		let current_funds = Balances::free_balance(AccountId32::new(ALICE));
 		let transfer_amount = 1;
 
 		add_task_to_task_queue(
 			ALICE,
 			vec![40],
 			vec![SCHEDULED_TIME],
-			Action::NativeTransfer { sender: AccountId32::new([ALICE; 32]), recipient: AccountId32::new([BOB; 32]), amount: transfer_amount },
+			Action::NativeTransfer { sender: AccountId32::new(ALICE), recipient: AccountId32::new(BOB), amount: transfer_amount },
 		);
 		add_task_to_task_queue(
 			ALICE,
 			vec![50],
 			vec![SCHEDULED_TIME],
-			Action::NativeTransfer { sender: AccountId32::new([ALICE; 32]), recipient: AccountId32::new([BOB; 32]), amount: transfer_amount },
+			Action::NativeTransfer { sender: AccountId32::new(ALICE), recipient: AccountId32::new(BOB), amount: transfer_amount },
 		);
 
 		LastTimeSlot::<Test>::put((LAST_BLOCK_TIME, LAST_BLOCK_TIME));
@@ -1146,8 +1146,8 @@ fn trigger_tasks_completes_some_native_transfer_tasks() {
 
 		AutomationTime::trigger_tasks(120_000);
 
-		assert_eq!(Balances::free_balance(AccountId32::new([ALICE; 32])), current_funds - (transfer_amount * 2));
-		assert_eq!(Balances::free_balance(AccountId32::new([BOB; 32])), transfer_amount * 2);
+		assert_eq!(Balances::free_balance(AccountId32::new(ALICE)), current_funds - (transfer_amount * 2));
+		assert_eq!(Balances::free_balance(AccountId32::new(BOB)), transfer_amount * 2);
 	})
 }
 
@@ -1175,7 +1175,7 @@ fn trigger_tasks_completes_some_xcmp_tasks() {
 			vec![(
 				(1, Junction::Parachain(PARA_ID.into())).into(),
 				Xcm(vec![Transact {
-					origin_type: OriginKind::SovereignAccount,
+					origin_type: OriginKind::Native,
 					require_weight_at_most: 100_000,
 					call: vec![3, 4, 5].into(),
 				}]),
@@ -1200,7 +1200,7 @@ fn trigger_tasks_xcmp_sends_error_event() {
 			vec![SCHEDULED_TIME],
 			Action::XCMP {
 				para_id: PARA_ID.try_into().unwrap(),
-				call: vec![9, 9, 9], // mock send_xcm will throw an error if call equals vec![9,9,9]
+				call: vec![9, 1, 1], // mocked send_xcm will throw an error if call equals vec![9,1,1]
 				weight_at_most: 100_000,
 			},
 		);
@@ -1335,7 +1335,7 @@ fn on_init_runs_tasks() {
 		assert_eq!(
 			events(),
 			[Event::AutomationTime(crate::Event::TaskMissed {
-				who: AccountId32::new([ALICE; 32]),
+				who: AccountId32::new(ALICE),
 				task_id: task_id3,
 				execution_time: LAST_BLOCK_TIME
 			})],
@@ -1385,12 +1385,12 @@ fn on_init_check_task_queue() {
 			events(),
 			[
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: tasks[3],
 					execution_time: LAST_BLOCK_TIME
 				}),
 				Event::AutomationTime(crate::Event::TaskMissed {
-					who: AccountId32::new([ALICE; 32]),
+					who: AccountId32::new(ALICE),
 					task_id: tasks[4],
 					execution_time: LAST_BLOCK_TIME
 				}),
@@ -1500,16 +1500,16 @@ fn migration_v2() {
 }
 
 fn schedule_task(
-	owner: u8,
+	owner: [u8; 32],
 	provided_id: Vec<u8>,
 	scheduled_times: Vec<u64>,
 	message: Vec<u8>,
 ) -> sp_core::H256 {
-	get_funds(AccountId32::new([owner; 32]));
+	get_funds(AccountId32::new(owner));
 	let task_hash_input =
-		TaskHashInput::<Test>::create_hash_input(AccountId32::new([owner; 32]), provided_id.clone());
+		TaskHashInput::<Test>::create_hash_input(AccountId32::new(owner), provided_id.clone());
 	assert_ok!(AutomationTime::schedule_notify_task(
-		Origin::signed(AccountId32::new([owner; 32])),
+		Origin::signed(AccountId32::new(owner)),
 		provided_id,
 		scheduled_times,
 		message,
@@ -1518,7 +1518,7 @@ fn schedule_task(
 }
 
 fn add_task_to_task_queue(
-	owner: u8,
+	owner: [u8; 32],
 	provided_id: Vec<u8>,
 	scheduled_times: Vec<u64>,
 	action: Action<Test>,
@@ -1531,7 +1531,7 @@ fn add_task_to_task_queue(
 }
 
 fn add_task_to_missed_queue(
-	owner: u8,
+	owner: [u8; 32],
 	provided_id: Vec<u8>,
 	scheduled_times: Vec<u64>,
 	action: Action<Test>,
@@ -1545,33 +1545,33 @@ fn add_task_to_missed_queue(
 }
 
 fn create_task(
-	owner: u8,
+	owner: [u8; 32],
 	provided_id: Vec<u8>,
 	scheduled_times: Vec<u64>,
 	action: Action<Test>,
 ) -> sp_core::H256 {
 	let task_hash_input = TaskHashInput::<Test>::create_hash_input(
-		AccountId32::new([owner; 32]),
+		AccountId32::new(owner),
 		provided_id.clone(),
 	);
 	let task_id = BlakeTwo256::hash_of(&task_hash_input);
 	let task = match action {
 		Action::Notify { message } => Task::<Test>::create_event_task(
-			AccountId32::new([owner; 32]),
+			AccountId32::new(owner),
 			provided_id,
 			scheduled_times.try_into().unwrap(),
 			message,
 		),
 		Action::NativeTransfer { sender: _, recipient, amount } =>
 			Task::<Test>::create_native_transfer_task(
-				AccountId32::new([owner; 32]),
+				AccountId32::new(owner),
 				provided_id,
 				scheduled_times.try_into().unwrap(),
 				recipient,
 				amount,
 			),
 		Action::XCMP { para_id, call, weight_at_most } => Task::<Test>::create_xcmp_task(
-			AccountId32::new([owner; 32]),
+			AccountId32::new(owner),
 			provided_id,
 			scheduled_times.try_into().unwrap(),
 			para_id,